LayoutTests/svg/animations/change-css-property-while-animating-fill-remove-expected.txt

SVG 1.1 dynamic animation tests


This tests scripting a CSS property while animation is running

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS parseFloat(getComputedStyle(rect).opacity) is 0
PASS parseFloat(getComputedStyle(rect).opacity) is 0.25
PASS parseFloat(getComputedStyle(rect).opacity) is 0.25
PASS parseFloat(getComputedStyle(rect).opacity) is 0.5
PASS parseFloat(getComputedStyle(rect).opacity) is 1
PASS parseFloat(getComputedStyle(rect).opacity) is 1
PASS successfullyParsed is true

TEST COMPLETE
